GAE Script for data replication
This commit introduces a simple GAE application that provides an endpoint,
namely '/recent_data', which allows the replication server to fetch the
data for a specified kind and for the required period of time.

Note that the app is not an off-the-shelf piece of software; it still has
to be configured with information about the kinds available for download
from Datastore, as well as some information about the structure of these kinds.